refactor(fiddle): extract ImgproxyControls.svelte (#254)
Moves the imgproxy tool sections (Resize through Metadata & color) from
App.svelte into a new ImgproxyControls.svelte child component, bound via
bind:fiddleState. The shared Request section (source + signature
controls), preview workspace, command bar, and signing flow remain in App.
Pure behavior-preserving refactor — no functionality change.
